+++ packages/xray-proton/src/parsers/burn.ts
@@ -2,68 +2,42 @@ import { LAMPORTS_PER_SOL } from "@solana/web3.js";
 
 import type { EnrichedTransaction, Source } from "helius-sdk";
 
-import { ProtonTransaction, ProtonTransactionAction } from "../types";
+import {
+    ProtonActionType,
+    ProtonTransaction,
+    ProtonTransactionAction,
+} from "../types";
 
-import { getSolanaName } from "@helius-labs/helius-namor";
+import { traverseTokenTransfers } from "../utils/token-transfers";
 
 export const parseBurn = (
     transaction: EnrichedTransaction
 ): ProtonTransaction => {
-    let source = "SYSTEM_PROGRAM" as Source;
+    const { tokenTransfers, signature, timestamp, type, source } = transaction;
+    const fee = transaction.fee / LAMPORTS_PER_SOL;
 
-    if (transaction?.tokenTransfers === null) {
+    if (tokenTransfers === null) {
         return {
             actions: [],
-            fee: 0,
+            fee,
             primaryUser: "",
-            signature: "",
+            signature,
             source,
-            timestamp: 0,
+            timestamp,
             type: "BURN",
         };
     }
 
-    const { tokenTransfers } = transaction;
+    const primaryUser = tokenTransfers[0].fromUserAccount || "";
     const actions: ProtonTransactionAction[] = [];
 
-    const primaryUser = tokenTransfers[0].fromUserAccount || "";
-
-    const { signature, timestamp, type } = transaction;
-    const fee = transaction.fee / LAMPORTS_PER_SOL;
-
-    source = transaction.source;
-
-    for (let i = 0; i < tokenTransfers.length; i++) {
-        const tx = tokenTransfers[i];
-        const from = tx.fromUserAccount || "";
-        let fromName;
-
-        if (tx.fromUserAccount) {
-            fromName = getSolanaName(tx.fromUserAccount);
-        }
-
-        const to = tx.toUserAccount || "";
-        let toName;
-
-        if (tx.toUserAccount) {
-            fromName = getSolanaName(tx.toUserAccount);
-        }
-
-        const sent = tx.mint;
-
-        // @ts-ignore
-        const amount = tx.tokenAmount;
-
-        actions.push({
-            actionType: "BURN",
-            amount,
-            from,
-            fromName,
-            sent,
-            to,
-            toName,
-        });
-    }
+    const address = undefined;
+    traverseTokenTransfers(
+        tokenTransfers,
+        actions,
+        address,
+        type as ProtonActionType
+    );
 
     return {
         actions,
