+++ packages/xray-proton/src/parsers/token.ts
@@ -1,12 +1,8 @@
-import { getSolanaName } from "@helius-labs/helius-namor";
 import { LAMPORTS_PER_SOL } from "@solana/web3.js";
 import { EnrichedTransaction, TokenTransfer } from "helius-sdk";
 import { ProtonParser, ProtonTransactionAction, SOL } from "../types";
 import { traverseNativeTransfers } from "../utils/native-transfers";
-
-interface TempTokenTransfer extends TokenTransfer {
-    tokenAmount: number;
-}
+import { traverseTokenTransfers } from "../utils/token-transfers";
 
 export const parseTokenMint: ProtonParser = (
     transaction: EnrichedTransaction,
@@ -38,68 +34,7 @@ export const parseTokenMint: ProtonParser = (
     const primaryUser = nativeTransfers[0]?.fromUserAccount || "";
     const actions: ProtonTransactionAction[] = [];
 
-    for (let i = 0; i < tokenTransfers.length; i++) {
-        const tx = tokenTransfers[i] as TempTokenTransfer;
-
-        const from = tx.fromUserAccount || "";
-        let fromName;
-        if (tx.fromUserAccount) {
-            fromName = getSolanaName(tx.fromUserAccount);
-        }
-
-        const to = tx.toUserAccount || "";
-        let toName;
-        if (tx.toUserAccount) {
-            toName = getSolanaName(tx.toUserAccount);
-        }
-
-        const amount = tx?.tokenAmount;
-
-        if (!address) {
-            const sent = tx.mint;
-            actions.push({
-                actionType: "TRANSFER",
-                amount,
-                from,
-                fromName,
-                sent,
-                to,
-                toName,
-            });
-        } else {
-            let actionType = "";
-            if (tx.fromUserAccount === address) {
-                actionType = "SENT";
-            } else if (tx.toUserAccount === address) {
-                actionType = "RECEIVED";
-            }
-
-            if (actionType === "SENT") {
-                const sent = tx.mint;
-                actions.push({
-                    actionType,
-                    amount,
-                    from,
-                    fromName,
-                    sent,
-                    to,
-                    toName,
-                });
-            } else if (actionType === "RECEIVED") {
-                const received = tx.mint;
-                actions.push({
-                    actionType,
-                    amount,
-                    from,
-                    fromName,
-                    received,
-                    to,
-                    toName,
-                });
-            }
-        }
-    }
-
+    traverseTokenTransfers(tokenTransfers, actions, address);
     traverseNativeTransfers(nativeTransfers, actions, address);
 
     return {
