+++ packages/xray-proton/src/parsers/transfer.ts
@@ -5,6 +5,7 @@ import { ProtonTransaction, ProtonTransactionAction, SOL } from "../types";
 import { getSolanaName } from "@helius-labs/helius-namor";
 import { traverseNativeTransfers } from "../utils/native-transfers";
 import { rentTransferCheck } from "../utils/rent-transfer-check";
+import { traverseTokenTransfers } from "../utils/token-transfers";
 
 interface TempTokenTransfer extends TokenTransfer {
     tokenAmount: number;
@@ -40,68 +41,7 @@ export const parseTransfer = (
     const primaryUser = tokenTransfers[0]?.fromUserAccount || "";
     const actions: ProtonTransactionAction[] = [];
 
-    for (let i = 0; i < tokenTransfers.length; i++) {
-        const tx = tokenTransfers[i] as TempTokenTransfer;
-
-        const from = tx.fromUserAccount || "";
-        let fromName;
-        if (tx.fromUserAccount) {
-            fromName = getSolanaName(tx.fromUserAccount);
-        }
-
-        const to = tx.toUserAccount || "";
-        let toName;
-        if (tx.toUserAccount) {
-            toName = getSolanaName(tx.toUserAccount);
-        }
-
-        const amount = tx?.tokenAmount;
-
-        if (!address) {
-            const sent = tx.mint;
-            actions.push({
-                actionType: "TRANSFER",
-                amount,
-                from,
-                fromName,
-                sent,
-                to,
-                toName,
-            });
-        } else {
-            let actionType = "";
-            if (tx.fromUserAccount === address) {
-                actionType = "TRANSFER_SENT";
-            } else if (tx.toUserAccount === address) {
-                actionType = "TRANSFER_RECEIVED";
-            }
-
-            if (actionType === "TRANSFER_SENT") {
-                const sent = tx.mint;
-                actions.push({
-                    actionType,
-                    amount,
-                    from,
-                    fromName,
-                    sent,
-                    to,
-                    toName,
-                });
-            } else if (actionType === "TRANSFER_RECEIVED") {
-                const received = tx.mint;
-                actions.push({
-                    actionType,
-                    amount,
-                    from,
-                    fromName,
-                    received,
-                    to,
-                    toName,
-                });
-            }
-        }
-    }
-
+    traverseTokenTransfers(tokenTransfers, actions, address);
     traverseNativeTransfers(nativeTransfers, actions, address);
 
     return {
