+++ packages/xray-proton/src/parsers/unknown.ts
@@ -2,13 +2,8 @@ import { LAMPORTS_PER_SOL } from "@solana/web3.js";
 import type { EnrichedTransaction, Source, TokenTransfer } from "helius-sdk";
 import { ProtonTransaction, ProtonTransactionAction, SOL } from "../types";
 
-import { getSolanaName } from "@helius-labs/helius-namor";
 import { traverseNativeTransfers } from "../utils/native-transfers";
-import { rentTransferCheck } from "../utils/rent-transfer-check";
-
-interface TempTokenTransfer extends TokenTransfer {
-    tokenAmount: number;
-}
+import { traverseTokenTransfers } from "../utils/token-transfers";
 
 export const parseUnknown = (
     transaction: EnrichedTransaction,
@@ -44,66 +39,7 @@ export const parseUnknown = (
 
     const actions: ProtonTransactionAction[] = [];
 
-    for (let i = 0; i < tokenTransfers.length; i++) {
-        const tx = tokenTransfers[i] as TempTokenTransfer;
-
-        const from = tx.fromUserAccount || "";
-        let fromName;
-        if (tx.fromUserAccount) {
-            fromName = getSolanaName(tx.fromUserAccount);
-        }
-
-        const to = tx.toUserAccount || "";
-        let toName;
-        if (tx.toUserAccount) {
-            toName = getSolanaName(tx.toUserAccount);
-        }
-
-        const amount = tx.tokenAmount;
-
-        if (!address) {
-            const sent = tx.mint;
-            actions.push({
-                actionType: "TRANSFER",
-                amount,
-                from,
-                fromName,
-                sent,
-                to,
-                toName,
-            });
-        } else {
-            const actionType =
-                address === tx.fromUserAccount
-                    ? "TRANSFER_SENT"
-                    : "TRANSFER_RECEIVED";
-
-            if (actionType === "TRANSFER_SENT") {
-                const sent = tx.mint;
-                actions.push({
-                    actionType,
-                    amount,
-                    from,
-                    fromName,
-                    sent,
-                    to,
-                    toName,
-                });
-            } else if (actionType === "TRANSFER_RECEIVED") {
-                const received = tx.mint;
-                actions.push({
-                    actionType,
-                    amount,
-                    from,
-                    fromName,
-                    received,
-                    to,
-                    toName,
-                });
-            }
-        }
-    }
-
+    traverseTokenTransfers(tokenTransfers, actions, address);
     traverseNativeTransfers(nativeTransfers, actions, address);
 
     return {
